7.1b version - running on master monitor
Needs fixing, merging etc. but working (except for 24 hour data
download email)

+++ hwconfig.h
@@ -6,8 +6,8 @@
 typedef struct {
    int8_t        OneWireGPIO;
    int8_t        ModBusSerial;
-   int8_t        ModBusBaudRate;
-   int8_t        ModBusSerialFormat;
+   int32_t       ModBusBaudRate;
+   int32_t        ModBusSerialFormat;
    int8_t        ModBusRxGPIO;
    int8_t        ModBusTxGPIO;
    int8_t        ModBusMsgDelay;
